test: make integration test layout vendor-first
google/ and microsoft/ each hold their mocks and messaging/calendar
suites; provider-agnostic helpers (http mock, entity store, connect,
channel queries) live in the shared utils folder.
