--- packages/twenty-server/test/integration/mocks/google/google-token-handlers.util.ts
@@ -1,22 +1,0 @@
-import { http, HttpResponse } from 'msw';
-
-import { GOOGLE_OAUTH_SCOPES } from 'test/integration/mocks/google/google-oauth-scopes.constant';
-import { type MswHandler } from 'test/integration/mocks/http-mock.util';
-
-const GOOGLE_TOKEN_URLS = [
-  'https://oauth2.googleapis.com/token',
-  'https://www.googleapis.com/oauth2/v4/token',
-];
-
-export const googleTokenHandlers = (): MswHandler[] =>
-  GOOGLE_TOKEN_URLS.map((url) =>
-    http.post(url, () =>
-      HttpResponse.json({
-        access_token: 'mock-access-token',
-        refresh_token: 'mock-refresh-token',
-        expires_in: 3600,
-        scope: GOOGLE_OAUTH_SCOPES,
-        token_type: 'Bearer',
-      }),
-    ),
-  );
