+++ packages/twenty-server/test/integration/utils/start-channel-sync-and-await.util.ts
@@ -1,4 +1,4 @@
-import { startChannelSync } from 'test/integration/messaging/utils/query-messaging.util';
+import { startChannelSync } from 'test/integration/utils/query-messaging.util';
 import { waitForAllJobsToFinish } from 'test/integration/utils/wait-for-all-jobs-to-finish.util';
 
 export const startChannelSyncAndAwait = async (
